eye-contact/requirements.txt
grpcio==1.62.2
grpcio-tools==1.62.2
flask==3.0.2

